+++ public/index.html
@@ -2,12 +2,12 @@
 	<head>
 		<title>Slay the Web</title>
 		<meta charset="utf-8" />
-		<link rel="stylesheet" href="index.css" />
+		<link rel="stylesheet" href="ui/index.css" />
 		<link href="https://fonts.googleapis.com/css2?family=Port Lligat Slab&amp;display=swap" rel="stylesheet">
 	</head>
 	<body>
 		<div id="root"></div>
 
-		<script async src="index.js" type="module"></script>
+		<script async src="ui/index.js" type="module"></script>
 	</body>
 </html>
