+++ public/ui/images/point.png
